+++ native/golang/types/ds/requests.go
@@ -42,6 +42,8 @@ type ReadDatasetRequest struct {
 	Encoding string `json:"encoding,omitempty"`
 	// Dataset name
 	Dsname string `json:"dsname"`
+	// Stream to write contents to
+	StreamId int `json:"stream,omitempty" tstype:"Writable"`
 }
 
 type WriteDatasetRequest struct {
@@ -55,6 +57,8 @@ type WriteDatasetRequest struct {
 	Dsname string `json:"dsname"`
 	// Dataset contents
 	Data string `json:"data" tstype:"B64String"`
+	// Stream to read contents from
+	StreamId int `json:"stream,omitempty" tstype:"Readable"`
 }
 
 type DeleteDatasetRequest struct {
@@ -71,16 +75,14 @@ type RestoreDatasetRequest struct {
 	Dsname string `json:"dsname"`
 }
 
-// default: DSORG=PO, RECFM=FB, LRECL=80
-// vb: DSORG=PO, RECFM=VB, LRECL=255
-// adata: DSORG=PO, RECFM=VB, LRECL=32756
 type CreateDatasetRequest struct {
 	common.CommandRequest `tstype:",extends"`
 	Command               string `json:"command" tstype:"\"createDataset\""`
 	// Dataset name
 	Dsname string `json:"dsname"`
-	// Type of the dataset to make
-	Type string `json:"dstype" tstype:"'default' | 'vb' | 'adata'"`
+
+	// Dataset attributes
+	Attributes common.DatasetAttributes `json:"attributes" tstype:"common.DatasetAttributes"`
 }
 
 type CreateMemberRequest struct {